make tests more reliable
The old tests check if the uploaded SVG was included in the HTML,
immediately after setting the HTML. In case Etherpad would remove the
image due to some error it's possible that the test returns true even if
the image gets removed in the end. To make this a little less likely,
prior to checking the HTML, let's wait for an ACCEPT message from the
server.